pyproject.toml
[project]
name = "rng-tui"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pyserial>=3.5",
    "pyusb>=1.3.1",
    "textual>=7.5.0",
]
